requirements.txt
bidict==0.22.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.6
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
